extract public types to ctypes module and rename nvlist to _nvlist
nvlist API is only used internally to libzfs_core.
Public interface of the latter uses a python dict instead.
So make it more obvious that nvlist is an internal detail.

Python types that represent C types of specific widths are public and thus
they are move to a module of their own.
Note that these types actually do not have to be CData types, instead
they could be custom python types that wrap a python integer value and
provide a method to cast to a corresponding CData type.