replace context managers nvlist_in, _out with a wrapper function nv_call
nv_call takes a function and a list of its arguments and makes a call.
The idea is that python dictionaries are used in place of all nvlist
in and out arguments in the arguments list. nv_call then examines
a signature of the function being called and automatically allocates
proper nvlist_t arguments and transfers data between the dictionaries
and the nvlists.