same FFI instance must be used for both libnvpair and libzfs_core bindings
The libraries are not independent, thus common types used by both libraries
must be declared in the same ffi instance.

Note that this change also requires that duplicate definitions are removed
from CDEFs of the binding modules. As a result a dependency order is
established between the libraries / modules.

While there, slightly re-organize __init__.py code, so that all private
definitions are placed within _setupCFFI() function.