lzc_snapshot: interface re-work and documentation
A list of snapshots now is just that, a list of snapshots.
Previously it used to be dictionary with all values being None.
This closely mimicked the C nvlist interface, but was an nuisance
in python.

Also, errlist is completely removed from the interface.
Any error details are embedded into an exception if thrown.

To do: clear up exception interface with respect possible multiple errord
To do: apply the same treatment to other interfaces that act on multiple
       entities