lazily call libzfs_core_init
One of benefits of the lazy initialization is that there is no access
to libzfs_core.so during documentation build or when setuptools are run.

At present not sure if ibzfs_core_init() and ibzfs_core_fini() should be
called explicitly or if there should be some smart mechanism of doing that
automatically.
The current approach is not ideal as it may leave /dev/zfs open even
if it's no longer needed.