rework how errors are reported for calls that operate on multiple entities
Now such a call would always raise a single specific exception subclassed
from MultipleOperationsFailure if an operation on at least one entity is
impossible to perform.
For example, in the case of lzc_snapshot() the SnapshotFailure will raised
if at least one snapshot is impossible to create regardless of an
actual number of requested snapshots, how many of them failed and for what
reasons.

MultipleOperationsFailure carries a list of more detailed error objects
with at least one element.  The error objects are instances of classes
derived from ZFSError.  An error object may contain a name of a failing
entity (filesystem, snapshot, etc) if it's possible to determine, which
is not always a case.