_TempPool; create requested filesystems before making pool read-only
Note that now the read-only import is done with -N flag, which instructs
to not mount any filesystems in the imported pool.  Without that we
would have to set an existing directory as a mount point for each
filesystem, because the mount points for child filesystems
can not be automatically created in read-only parent filesystems.