use a random, not predictable, name for a file used to modify a filesystem
Using tempfile.NamedTemporaryFile() ensures that the temporary file
is not left behind and that its existence and predictable name does not
cause interference between tests.

One shortcoming of the new approach is that now the file has to be
explicitly flush-ed to be sure that its content is in ZFS and thus we
can reason about the dataset's size based on the file's size.
We can not just close the file because in that case it it automatically
unlinked.