switch to different libzfs_core extensions
Some notes:

_uncommitted decorator now takes an optional argument that makes
a status of a wrapped function dependent on another function.
This is needed to wrap several functions that all are implemented
via lzc_list.

nvlist_unpack() from libnvpair is exposed to the Python code, so
that we can unpack nvlists that we read from a file descriptor
returned by lzc_list().

Constants are defined for layout and size of the pipe record header.
Using Python struct module to unpack those records is much easier
than employing cffi for that.

It would probably make sense to turn lzc_list() into a context
manager, so that the pipe file descriptors (for both ends) could
be closed automatically.