extract repeating patterns of name validity and length checks
Also fix bugs uncovered by the change.
Previously in some conditions an exception was returned instead of being
raised and, thus, it was ignored.  As a result behavior lzc_set_prop
and lzc_inherit was incorrect and it was incorrectly documented.
Additionally, dataset name validation was incorrect in lzc_set_prop and
lzc_get_props.

Fix the test cases and the documentation accordingly.