lzc_list_snaps can be misused / abused on a snapshot
That situation could be considered as an error, but it is not now.
It's documented that in that case an empty listing is produced.
But lzc_list() would produce a list that includes the snapshot
itself, so we have to ignore it similarly to what is done in
lzc_list_children (but for different reasons).