add illumos support for mounting ZFS filesystem and snapshots
First, 'mount -F zfs' is used on illumos as opposed to 'mount -t zfs'
on FreeBSD and Linux to mount a filesystem with a legacy mountpoint.

Second, a snapshot can not be explicitly mounted on illumos as
it is not considered to be an independent mountable entity.
There are two ways to access the snapshot's content:
- create a read-only clone filesystem and mount that
- traverse to the snapshot content through .zfs/snapshot/<name>

At the moment the latter mechanism is used on illumos.
In fact, both of the above mechanisms should work on all supported
platforms, so we might switch to one of them for all platforms
in the future.