lzc_get_props acts only on the specified dataset, so no recursion
Previously, lzc_get_props did not set 'recursion' argument
for _list() which resulted in the unlimited recursion and thus
we get properties of a child dataset instead of the requested one.

While there fix a problem where we passed 'type' as a boolean nvlist
element.  The kernel treated it the same as a missing element because
of the unexpected type, but that was fragile and confusing.
So, now we do not set 'type' element at all if no type is explicitly
requested.

Also, if 'recurse' parameter is set to zero, then we do not set
'recurse' element at all, which is a more natural indicator that
no recursion is asked for.

Finally, a test case is added to check for the fixed problem.